requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
darkdetect==0.8.0
idna==3.10
macholib==1.16.3
modulegraph==0.19.6
packaging==24.2
pillow==11.0.0
py2app==0.28.8
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pystray==0.19.5
requests==2.32.3
setuptools==75.6.0
six==1.17.0
urllib3==2.2.3
watchdog==6.0.0
